Dev (#2)
* Implimented get asks by categories, enabled strict type checking and fixed req/res types in controllers

* Added filters for querying Banned and Hidden users and asks respectively.

* Validate Id, applied sorting

* Modified  users and asks models

+++ controllers/askController.ts
@@ -1,5 +1,6 @@
 import Ask from "../models/askModel";
 import Express from "express";
+import mongoose from "mongoose";
 
 /** METHODS:
  * ask_all_get
@@ -15,25 +16,25 @@ const ask_all_get = (req: Express.Request, res: Express.Response) => {
   /** hide Hidden defaults to true */
   /// Only hidden
   if (queryOptions.onlyHidden) {
-    Ask.find({ "status.hidden": { $eq: true } })
+    Ask.find({ "status.hidden": { $eq: true } }).sort({createdAt: -1})
       .then(result => res.status(200).json(result))
       .catch(err => res.status(400).json({ message: err.message }));
   }
   /// All asks regardless
   else if (queryOptions.showHidden === true) {
-    Ask.find({})
+    Ask.find({}).sort({createdAt: -1})
       .then(result => res.status(200).json(result))
       .catch(err => res.status(400).json({ message: err.message }));
   }
   /// Categorized unhidden
   else if (queryOptions.categories !== undefined) {
-    Ask.find({ $and: [{ categories: { $in: queryOptions.categories } }, { "status.hidden": { $eq: false } }] })
+    Ask.find({ $and: [{ categories: { $in: queryOptions.categories } }, { "status.hidden": { $eq: false } }] }).sort({createdAt: -1})
       .then(result => res.status(200).json(result))
       .catch(err => res.status(400).json({ message: err.message }));
   }
   /// All unhidden
   else {
-    Ask.find({ "status.hidden": { $eq: false } })
+    Ask.find({ "status.hidden": { $eq: false } }).sort({createdAt: -1})
       .then(result => res.status(200).json(result))
       .catch(err => res.status(400).json({ message: err.message }));
   }
@@ -51,6 +52,9 @@ const ask_create_post = (req: Express.Request, res: Express.Response) => {
 /** Edit an existing ask on the database with entry */
 const ask_edit_patch = (req: Express.Request, res: Express.Response) => {
   const id = req.params.id;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({error: 'ask not found'});
+  }
   Ask.findByIdAndUpdate(id, req.body)
     .then(() => res.status(201).end())
     .catch(err => res.status(400).json({ message: err.message }));
@@ -59,6 +63,9 @@ const ask_edit_patch = (req: Express.Request, res: Express.Response) => {
 /** Deletes the ask with id provided to the req params */
 const ask_delete = (req: Express.Request, res: Express.Response) => {
   const id = req.params.id;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({error: 'ask not found'});
+  }
   Ask.findByIdAndDelete(id)
     .then(() => res.status(200).end())
     .catch(err => res.status(400).json({ message: err.message }));
@@ -66,6 +73,9 @@ const ask_delete = (req: Express.Request, res: Express.Response) => {
 /** Hide the ask with given id */
 const ask_hide_patch = (req: Express.Request, res: Express.Response) => {
   const id = req.params.id;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({error: 'ask not found'});
+  }
   Ask.findByIdAndUpdate(id, req.body)
     .then(() => res.status(200).end())
     .catch(err => res.status(400).json({ message: err.message }));
